refactor(tui): implement shell architecture with persistent menu/status bar
Replace MainWindow with TuiShell that provides:
- Persistent menu bar visible on all screens
- Persistent status bar (no duplication)
- Content area where screens swap via NavigateTo/NavigateBack
- Context-aware menus (screen-specific items inserted between File and Help)

Refactor SqlQueryScreen to implement ITuiScreen interface:
- No longer inherits from Window
- Provides Content view, Title, and ScreenMenuItems
- Uses CloseRequested event for navigation instead of RequestStop
- Hotkeys registered in OnActivated, unregistered in OnDeactivating

This fixes the issue where menu bar was not visible on SqlQueryScreen
and eliminates duplicated status bar code.
