+++ src/PPDS.Cli/Tui/Screens/ITuiScreen.cs
@@ -1,0 +1,46 @@
+using PPDS.Cli.Tui.Infrastructure;
+using Terminal.Gui;
+
+namespace PPDS.Cli.Tui.Screens;
+
+/// <summary>
+/// Interface for screens that can be hosted in the TuiShell.
+/// Screens provide content and optional screen-specific menus.
+/// </summary>
+internal interface ITuiScreen : IDisposable
+{
+    /// <summary>
+    /// The content view to display in the shell's content area.
+    /// </summary>
+    View Content { get; }
+
+    /// <summary>
+    /// The title for the content frame.
+    /// </summary>
+    string Title { get; }
+
+    /// <summary>
+    /// Optional screen-specific menu items (inserted between File and Help).
+    /// Return null if no screen-specific menus are needed.
+    /// </summary>
+    MenuBarItem[]? ScreenMenuItems { get; }
+
+    /// <summary>
+    /// Raised when the screen wants to close (e.g., user pressed Escape).
+    /// The shell subscribes to this and calls NavigateBack().
+    /// </summary>
+    event Action? CloseRequested;
+
+    /// <summary>
+    /// Called when the screen becomes active (after content is added to shell).
+    /// Register screen-scope hotkeys here.
+    /// </summary>
+    /// <param name="hotkeyRegistry">The hotkey registry to register screen-scope hotkeys with.</param>
+    void OnActivated(IHotkeyRegistry hotkeyRegistry);
+
+    /// <summary>
+    /// Called when the screen is about to be deactivated (before content removal).
+    /// Unregister screen-scope hotkeys here.
+    /// </summary>
+    void OnDeactivating();
+}
